pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-mcp-server"
version = "1.0.0"
description = "A local MCP server suite for Crypto trading, Portfolio tracking, and LLM reasoning."
authors = [{name = "Corax CoLAB", email = "info@coraxcolab.com"}]
dependencies = [
    "mcp[cli]>=1.8.0",
    "fastmcp>=0.3.0",
    "ccxt>=2.4.0",
    "web3>=7.0.0",
    "pandas>=2.0.0",
    "pandas-ta",
    "pycoingecko>=1.0.0",
    "uvicorn[standard]>=0.22.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "litellm>=1.0.0",
    "requests-cache>=0.9.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]
